requirements.txt
APScheduler==3.9.1
azure-common==1.1.28
azure-core==1.25.1
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Flask==2.2.2
Flask-APScheduler==1.12.4
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msrest==0.7.1
oauthlib==3.2.0
psycopg2==2.9.3
pycparser==2.21
pyodbc==4.0.34
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
